server/app/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
alembic==1.14.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
passlib==1.7.4
pydantic-settings==2.7.1
requests==2.32.3
PyJWT==2.4.0
pydantic[email]
